rickandmortyapi/requirements.txt
django
django-environ==0.11.2
psycopg[binary,pool]
requests>=2.32.1

# Django REST Framework
djangorestframework==3.15.2

# Json progression bar
tqdm==4.66.5
